Rename RequestSocketFrame and ResponseSocketFrame as this are misleading names
