+++ reformcloud2-executor-api/src/main/java/systems/reformcloud/reformcloud2/executor/api/http/websocket/response/ResponseSocketFrameFactory.java
@@ -50,5 +50,5 @@ public abstract class ResponseSocketFrameFactory {
      * @return the created response frame.
      */
     @NotNull
-    public abstract ResponseSocketFrame<?> forFrame(@NotNull SocketFrame<?> socketFrame);
+    public abstract ResponseFrameHolder<?> forFrame(@NotNull SocketFrame<?> socketFrame);
 }
